hoan thanh so bo entity

+++ src/main/java/com/example/project_pulse_backend/entity/User.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -17,31 +18,31 @@ import java.util.List;
 @Builder
 @JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
 public class User {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "id")
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
 
-	@Column(name = "user_name", nullable = false)
-	private String userName;
+    @Column(name = "user_name", nullable = false)
+    private String userName;
 
-	@Column(name = "position", nullable = false)
-	private String position;
+    @Column(name = "position", nullable = false)
+    private String position;
 
-	@Column(name = "resource_rate")
-	private float resourceRate;
+    @Column(name = "resource_rate")
+    private BigDecimal resourceRate;
 
-	@Column(name = "avatar_url")
-	private String avatarUrl;
+    @Column(name = "avatar_url")
+    private String avatarUrl;
 
-	@Column(name = "hidden")
-	private boolean hidden;
+    @Column(name = "hidden")
+    private boolean hidden;
 
-	@ManyToOne
-	@JoinColumn(name = "department_id", nullable = false)
-	private Department department;
+    @ManyToOne
+    @JoinColumn(name = "department_id", nullable = false)
+    private Department department;
 
-	@ManyToMany(fetch = FetchType.LAZY)
-	@JoinTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_name"))
-	private List<Role> roles;
+    @ManyToMany(fetch = FetchType.LAZY)
+    @JoinTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_name"))
+    private List<Role> roles;
 }
